20_OCRovai_Web_App/requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
click==7.1.2
flake8==3.8.2
flake8-polyfill==1.0.2
flake8-quotes==3.2.0
Flask==1.1.2
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
numpy==1.18.1
olefile==0.46
packaging==20.4
pep8-naming==0.10.0
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pyperclip==1.7.0
pytesseract==0.3.4
pytest==5.4.2
six==1.15.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
